Seperate Rendering code in seperate file
